track the sent emails in a txt file
